+++ src/components/common/EntityActions.tsx
@@ -1,63 +1,42 @@
-import { useNavigate } from 'react-router-dom';
+import { useNavigate } from "react-router-dom";
 
 interface EntityActionsProps {
-  id?: string | number;
-  onEdit?: () => void;
-  onDelete?: () => void;
+  id: number;
+  onDelete: () => void;
   detailsPath?: string;
   detailsLabel?: string;
   deleteIcon?: string;
-  showDetails?: boolean;
-  showEdit?: boolean;
-  showDelete?: boolean;
+  onDetails?: () => void; 
 }
 
 export const EntityActions = ({
-  onEdit,
+  id,
   onDelete,
   detailsPath,
   detailsLabel = "Ver Detalles",
-  deleteIcon = "🗑️",
-  showDetails = false,
-  showEdit = true,
-  showDelete = true,
+  deleteIcon,
+  onDetails,
 }: EntityActionsProps) => {
   const navigate = useNavigate();
 
-  const handleDetailsClick = () => {
-    if (detailsPath) {
-      navigate(detailsPath);
-    }
-  };
-
   return (
-    <div className="flex space-x-2">
-      {showDetails && (
-        <button
-          onClick={handleDetailsClick}
-          className="bg-purple-600 hover:bg-purple-700 text-white px-3 py-1 rounded-md text-sm transition-colors"
-        >
-          {detailsLabel}
-        </button>
-      )}
-      
-      {showEdit && onEdit && (
-        <button
-          onClick={onEdit}
-          className="bg-blue-600 hover:bg-blue-700 text-white px-3 py-1 rounded-md text-sm transition-colors"
-        >
-          Editar
-        </button>
-      )}
-      
-      {showDelete && onDelete && (
-        <button
-          onClick={onDelete}
-          className="bg-red-600 hover:bg-red-700 text-white px-3 py-1 rounded-md text-sm transition-colors"
-        >
-          {deleteIcon}
-        </button>
-      )}
+    <div className="flex gap-3 items-center justify-center">
+      <button
+        className="bg-purple-600 text-white px-6 py-2 rounded-full font-semibold hover:bg-purple-700 transition cursor-pointer"
+        onClick={onDetails ? onDetails : () => detailsPath && navigate(detailsPath)}
+      >
+        {detailsLabel}
+      </button>
+      <button
+        className="rounded-full p-1 hover:bg-red-100 transition flex items-center justify-center cursor-pointer"
+        onClick={onDelete}
+      >
+        {deleteIcon ? (
+          <img src={deleteIcon} alt="Eliminar" className="w-10 h-10" />
+        ) : (
+          <span role="img" aria-label="Eliminar" className="text-2xl">🗑️</span>
+        )}
+      </button>
     </div>
   );
 };
